third_party/blink/web_tests/editing/inserting/insert-paragraph-separator-in-paragraph-in-list-expected.txt

insert at end
PASS actualHTML is "<li><p><br></p></li>"
PASS selection.type is "Caret"
PASS selection.focusNode.textContent is ""
PASS selection.focusOffset is 0


insert at start
PASS actualHTML is "<li><p><br></p></li>"
PASS selection.type is "Caret"
PASS selection.focusNode.textContent is "is"
PASS selection.focusOffset is 0


insert at middle
PASS actualHTML is "<li><p>his</p></li>"
PASS selection.type is "Caret"
PASS selection.focusNode.textContent is "his"
PASS selection.focusOffset is 0


PASS successfullyParsed is true

TEST COMPLETE
